float to half

+++ PowerPBS/Lib/BSDF.hlsl
@@ -2,117 +2,117 @@
 #define BSDF_HLSL
 #include "CommonUtils.hlsl"
 
-inline float FastSSS(float3 l,float3 v){
+inline half FastSSS(half3 l,half3 v){
     return saturate(dot(l,v));
 }
 
-inline float DisneyDiffuse(float nv,float nl,float lh,float roughness){
-    float fd90 = 0.5 + 2*roughness*lh*lh;
-    float lightScatter = 1 - (fd90 - 1) * Pow5(1 - nl);
-    float viewScatter = 1 - (fd90 - 1 ) * Pow5(1 - nv);
+inline half DisneyDiffuse(half nv,half nl,half lh,half roughness){
+    half fd90 = 0.5 + 2*roughness*lh*lh;
+    half lightScatter = 1 - (fd90 - 1) * Pow5(1 - nl);
+    half viewScatter = 1 - (fd90 - 1 ) * Pow5(1 - nv);
     return lightScatter * viewScatter;
 }
-inline float RoughnessToSpecPower(float a){
-    float a2 = a * a;
-    float sq = max(1e-4f,a2 * a2);
-    float n = 2.0/sq - 2;
+inline half RoughnessToSpecPower(half a){
+    half a2 = a * a;
+    half sq = max(1e-4f,a2 * a2);
+    half n = 2.0/sq - 2;
     n = max(n,1e-4f);
     return n;
 }
-inline float3 FresnelTerm(float3 F0,float lh){
+inline half3 FresnelTerm(half3 F0,half lh){
     return F0 + (1-F0) * Pow5(1 - lh);
 }
-inline float3 FresnelLerp(float3 f0,float3 f90,float lh){
-    float t = Pow5(1-lh);
+inline half3 FresnelLerp(half3 f0,half3 f90,half lh){
+    half t = Pow5(1-lh);
     return lerp(f0,f90,t);
 }
-inline float3 FresnelLerpFast(float3 F0,float3 F90,float lh){
-    float t = Pow4(1 - lh);
+inline half3 FresnelLerpFast(half3 F0,half3 F90,half lh){
+    half t = Pow4(1 - lh);
     return lerp(F0,F90,t);
 }
 
-inline float SmithJointGGXTerm(float nl,float nv,float a2){
-    float v = nv * (nv * (1-a2)+a2);
-    float l = nl * (nl * (1-a2)+a2);
+inline half SmithJointGGXTerm(half nl,half nv,half a2){
+    half v = nv * (nv * (1-a2)+a2);
+    half l = nl * (nl * (1-a2)+a2);
     return 0.5f/(v + l + 1e-5f);
 }
 
-inline float NDFBlinnPhongTerm(float nh,float a){
-    float normTerm = (a + 2)* 0.5/PI;
-    float specularTerm = pow(nh,a);
+inline half NDFBlinnPhongTerm(half nh,half a){
+    half normTerm = (a + 2)* 0.5/PI;
+    half specularTerm = pow(nh,a);
     return normTerm * specularTerm;
 }
 
-inline float D_GGXTerm(float nh,float a2){
-    float d = (nh*a2-nh)*nh + 1;
+inline half D_GGXTerm(half nh,half a2){
+    half d = (nh*a2-nh)*nh + 1;
     return a2 / (d*d + 1e-7f);
 }
 
 
-inline float D_GGXAnisoNoPI(float TdotH, float BdotH, float NdotH, float roughnessT, float roughnessB)
+inline half D_GGXAnisoNoPI(half TdotH, half BdotH, half NdotH, half roughnessT, half roughnessB)
 {
-    float a2 = roughnessT * roughnessB;
-    float3 v = float3(roughnessB * TdotH, roughnessT * BdotH, a2 * NdotH);
-    float  s = dot(v, v);
+    half a2 = roughnessT * roughnessB;
+    half3 v = half3(roughnessB * TdotH, roughnessT * BdotH, a2 * NdotH);
+    half  s = dot(v, v);
 
     // If roughness is 0, returns (NdotH == 1 ? 1 : 0).
     // That is, it returns 1 for perfect mirror reflection, and 0 otherwise.
     return SafeDiv(a2 * a2 * a2, s * s);
 }
 
-float D_GGXAniso(float TdotH, float BdotH, float NdotH, float roughnessT, float roughnessB)
+half D_GGXAniso(half TdotH, half BdotH, half NdotH, half roughnessT, half roughnessB)
 {
     return INV_PI * D_GGXAnisoNoPI(TdotH, BdotH, NdotH, roughnessT, roughnessB);
 }
 
-float D_WardAniso(float nl,float nv,float nh,float th,float bh,float roughT,float roughB){
-    float roughTH = th/roughT;
-    float roughBH = bh/roughB;
+half D_WardAniso(half nl,half nv,half nh,half th,half bh,half roughT,half roughB){
+    half roughTH = th/roughT;
+    half roughBH = bh/roughB;
     return sqrt(max(0,nl/nv)) * exp(-2 * (roughTH*roughTH+roughBH*roughBH)/(1+nh));
 }
 
-float BankBRDF(float3 l,float3 v,float3 t,float ks,float power){
-    float lt = dot(l,t);
-    float vt = dot(v,t);
-    float lt2 = lt*lt;
-    float vt2 = vt*vt;
+half BankBRDF(half3 l,half3 v,half3 t,half ks,half power){
+    half lt = dot(l,t);
+    half vt = dot(v,t);
+    half lt2 = lt*lt;
+    half vt2 = vt*vt;
     return ks * pow(sqrt(1-lt2)*sqrt(1-vt2) - lt*vt,power);
 }
 
-float GetSmithJointGGXAnisoPartLambdaV(float TdotV, float BdotV, float NdotV, float roughnessT, float roughnessB)
+half GetSmithJointGGXAnisoPartLambdaV(half TdotV, half BdotV, half NdotV, half roughnessT, half roughnessB)
 {
-    return length(float3(roughnessT * TdotV, roughnessB * BdotV, NdotV));
+    return length(half3(roughnessT * TdotV, roughnessB * BdotV, NdotV));
 }
 
 // Note: V = G / (4 * NdotL * NdotV)
 // Ref: https://cedec.cesa.or.jp/2015/session/ENG/14698.html The Rendering Materials of Far Cry 4
-float V_SmithJointGGXAniso(float TdotV, float BdotV, float NdotV, float TdotL, float BdotL, float NdotL, float roughnessT, float roughnessB, float partLambdaV)
+half V_SmithJointGGXAniso(half TdotV, half BdotV, half NdotV, half TdotL, half BdotL, half NdotL, half roughnessT, half roughnessB, half partLambdaV)
 {
-    float lambdaV = NdotL * partLambdaV;
-    float lambdaL = NdotV * length(float3(roughnessT * TdotL, roughnessB * BdotL, NdotL));
+    half lambdaV = NdotL * partLambdaV;
+    half lambdaL = NdotV * length(half3(roughnessT * TdotL, roughnessB * BdotL, NdotL));
 
     return 0.5 / (lambdaV + lambdaL);
 }
 
-float V_SmithJointGGXAniso(float TdotV, float BdotV, float NdotV, float TdotL, float BdotL, float NdotL, float roughnessT, float roughnessB)
+half V_SmithJointGGXAniso(half TdotV, half BdotV, half NdotV, half TdotL, half BdotL, half NdotL, half roughnessT, half roughnessB)
 {
-    float partLambdaV = GetSmithJointGGXAnisoPartLambdaV(TdotV, BdotV, NdotV, roughnessT, roughnessB);
+    half partLambdaV = GetSmithJointGGXAnisoPartLambdaV(TdotV, BdotV, NdotV, roughnessT, roughnessB);
     return V_SmithJointGGXAniso(TdotV, BdotV, NdotV, TdotL, BdotL, NdotL, roughnessT, roughnessB, partLambdaV);
 }
 // Inline D_GGXAniso() * V_SmithJointGGXAniso() together for better code generation.
-float DV_SmithJointGGXAniso(float TdotH, float BdotH, float NdotH, float NdotV,
-                           float TdotL, float BdotL, float NdotL,
-                           float roughnessT, float roughnessB, float partLambdaV)
+half DV_SmithJointGGXAniso(half TdotH, half BdotH, half NdotH, half NdotV,
+                           half TdotL, half BdotL, half NdotL,
+                           half roughnessT, half roughnessB, half partLambdaV)
 {
-    float a2 = roughnessT * roughnessB;
-    float3 v = float3(roughnessB * TdotH, roughnessT * BdotH, a2 * NdotH);
-    float  s = dot(v, v);
+    half a2 = roughnessT * roughnessB;
+    half3 v = half3(roughnessB * TdotH, roughnessT * BdotH, a2 * NdotH);
+    half  s = dot(v, v);
 
-    float lambdaV = NdotL * partLambdaV;
-    float lambdaL = NdotV * length(float3(roughnessT * TdotL, roughnessB * BdotL, NdotL));
+    half lambdaV = NdotL * partLambdaV;
+    half lambdaL = NdotV * length(half3(roughnessT * TdotL, roughnessB * BdotL, NdotL));
 
-    float2 D = float2(a2 * a2 * a2, s * s);  // Fraction without the multiplier (1/Pi)
-    float2 G = float2(1, lambdaV + lambdaL); // Fraction without the multiplier (1/2)
+    half2 D = half2(a2 * a2 * a2, s * s);  // Fraction without the multiplier (1/Pi)
+    half2 G = half2(1, lambdaV + lambdaL); // Fraction without the multiplier (1/2)
 
     // This function is only used for direct lighting.
     // If roughness is 0, the probability of hitting a punctual or directional light is also 0.
@@ -120,44 +120,44 @@ float DV_SmithJointGGXAniso(float TdotH, float BdotH, float NdotH, float NdotV,
     return (INV_PI * 0.5) * (D.x * G.x) / max(D.y * G.y, FLT_MIN);
 }
 
-float DV_SmithJointGGXAniso(float TdotH, float BdotH, float NdotH,
-                           float TdotV, float BdotV, float NdotV,
-                           float TdotL, float BdotL, float NdotL,
-                           float roughnessT, float roughnessB)
+half DV_SmithJointGGXAniso(half TdotH, half BdotH, half NdotH,
+                           half TdotV, half BdotV, half NdotV,
+                           half TdotL, half BdotL, half NdotL,
+                           half roughnessT, half roughnessB)
 {
-    float partLambdaV = GetSmithJointGGXAnisoPartLambdaV(TdotV, BdotV, NdotV, roughnessT, roughnessB);
+    half partLambdaV = GetSmithJointGGXAnisoPartLambdaV(TdotV, BdotV, NdotV, roughnessT, roughnessB);
     return DV_SmithJointGGXAniso(TdotH, BdotH, NdotH, NdotV,
                                  TdotL, BdotL, NdotL,
                                  roughnessT, roughnessB, partLambdaV);
 }
 
-float CharlieD(float roughness, float ndoth)
+half CharlieD(half roughness, half ndoth)
 {
-    float invR = 1. / roughness;
-    float cos2h = ndoth * ndoth;
-    float sin2h = max(1. - cos2h,0.0078125);
+    half invR = 1. / roughness;
+    half cos2h = ndoth * ndoth;
+    half sin2h = max(1. - cos2h,0.0078125);
     return (2. + invR) * pow(sin2h, invR * .5) / (2. * PI);
 }
 
-float AshikhminV(float ndotv, float ndotl)
+half AshikhminV(half ndotv, half ndotl)
 {
     return 1. / (4. * (ndotl + ndotv - ndotl * ndotv));
 }
 
-float D_Ashikhmin(float roughness, float NoH) {
+half D_Ashikhmin(half roughness, half NoH) {
     // Ashikhmin 2007, "Distribution-based BRDFs"
-	float a2 = roughness * roughness;
-	float cos2h = NoH * NoH;
-	float sin2h = max(1.0 - cos2h, 0.0078125); // 2^(-14/2), so sin2h^2 > 0 in fp16
-	float sin4h = sin2h * sin2h;
-	float cot2 = -cos2h / (a2 * sin2h);
+	half a2 = roughness * roughness;
+	half cos2h = NoH * NoH;
+	half sin2h = max(1.0 - cos2h, 0.0078125); // 2^(-14/2), so sin2h^2 > 0 in fp16
+	half sin4h = sin2h * sin2h;
+	half cot2 = -cos2h / (a2 * sin2h);
 	return 1.0 / (PI * (4.0 * a2 + 1.0) * sin4h) * (4.0 * exp(cot2) + sin4h);
 }
 
-float MinimalistCookTorrance(float nh,float lh,float rough,float rough2){
-    float d = nh * nh * (rough2-1) + 1.00001f;
-    float lh2 = lh * lh;
-    float spec = rough2/((d*d) * max(0.1,lh2) * (rough*4+2)); // approach sqrt(rough2)
+half MinimalistCookTorrance(half nh,half lh,half rough,half rough2){
+    half d = nh * nh * (rough2-1) + 1.00001f;
+    half lh2 = lh * lh;
+    half spec = rough2/((d*d) * max(0.1,lh2) * (rough*4+2)); // approach sqrt(rough2)
     
     #if defined (SHADER_API_MOBILE) || defined (SHADER_API_SWITCH)
         spec = clamp(spec,0,100);
